save for today

+++ mysql.sql
@@ -3,7 +3,7 @@ CREATE TABLE `bankaccounts` (
   `accountname` VARCHAR( 255 ) NOT NULL ,
   `players` LONGTEXT NOT NULL,
   `password` VARCHAR( 255 ) NULL DEFAULT '',
-  `amount` DOUBLE NOT NULL DEFAULT '0'
+  `amount` DOUBLE( 255,2 ) NOT NULL DEFAULT '0'
 )
 CREATE TABLE `bankareas` (
   `id` INT( 255 ) NOT NULL AUTO_INCREMENT PRIMARY KEY ,
